moved screens related classes and enums into Menu package
