+++ src/main/java/inf112/BalmersPeak/app/GUI.java
@@ -4,6 +4,7 @@ import com.badlogic.gdx.Game;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL30;
+import inf112.balmerspeak.app.Menu.MenuScreen;
 
 
 public class GUI extends Game {
