+++ src/main/java/inf112/BalmersPeak/app/Menu/MenuScreen.java
@@ -1,16 +1,13 @@
-package inf112.balmerspeak.app;
+package inf112.balmerspeak.app.Menu;
 
-import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.audio.Sound;
-import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
-
+import inf112.balmerspeak.app.GUI;
 
 public class MenuScreen extends MainScreen implements Screen {
     // Game object
