+++ src/main/java/inf112/BalmersPeak/app/Menu/RulesScreen.java
@@ -1,15 +1,9 @@
-package inf112.balmerspeak.app;
+package inf112.balmerspeak.app.Menu;
 
-import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.scenes.scene2d.Actor;
-import com.badlogic.gdx.scenes.scene2d.InputEvent;
-import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
-import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
-import com.badlogic.gdx.utils.viewport.ScreenViewport;
+import inf112.balmerspeak.app.GUI;
 
 public class RulesScreen extends MainScreen implements Screen {
 
@@ -28,6 +22,9 @@ public class RulesScreen extends MainScreen implements Screen {
 
         // Get root table
         Table root = super.getRoot();
+
+        // Get skin
+        Skin skin = super.getSkin();
 
         // Get button label
         Label backbtnLabel = super.getBtnLabel("Back");
